+++ docs/CHANGELOG.md
@@ -7,6 +7,11 @@ All notable changes to *Quality-time* will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- The line "## <square-bracket>Unreleased</square-bracket>" is replaced by the release/release.py script with the new release version and release date. -->
+## [Unreleased]
+
+### Added
+
+- Add 'Missing metrics' metric. Closes [#1477](https://github.com/ICTU/quality-time/issues/1477).
 
 ## [3.22.0] - [2021-05-26]